Remove all empty directories from $HOME/Desktop
By piping the results of the `find` command to `tac`, we get the results in reverse, like so:
```
/User/dfelton/foo/bar
/User/dfelton/foo
```
This allows us to then essentially "recursively remove empty directories", as once all sub-directories of a directory are removed, if THAT directory is also now empty, it gets removed as well.

Additional check added to ensure we don't actually delete our desktop.

Enjoy your clean(er) desktop :)

Update desktop-cleaner.bash